updated for 64 bit; removed depricated nsstring things

+++ src/models/src/NICellCatalog.m
@@ -83,7 +83,7 @@
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 - (id)initWithStyle:(UITableViewCellStyle)style reuseIdentifier:(NSString *)reuseIdentifier {
-  if ((self = [super initWithStyle:style reuseIdentifier:[reuseIdentifier stringByAppendingFormat:@"%d", style]])) {
+  if ((self = [super initWithStyle:style reuseIdentifier:[reuseIdentifier stringByAppendingFormat:@"%li", (long)style]])) {
     self.selectionStyle = UITableViewCellSelectionStyleNone;
 	  [self.detailTextLabel setNumberOfLines:2];
   }
